Consolidation of two previous commits: AbaqusInputGeneration will now generate either polygonal or half-infinite space depending on what is specified in the input.yml file. Note that yml files cannot be immediately run from within "abaqus python" instance
To allow this on Windows, make sure that a Python instance of the same version as Abaqus' version is installed (at the time of commit, this is 2.7.3), and install the required modules onto this (yaml). At the top of each script which will be run by "abaqus python", do "sys.path.append(r"C:\Path-to-python-2.7.3-site-packages")". On this install, the path is "C:\Python27\Lib\site-packages".
On Linux (tested on BluePebble), upload the relevent modules (i.e. the entire yaml folder within site-packages) to the cluster from the Windows installation. At the top of each python script, do "sys.path.append(r"/user/home/<user>/path-to-package-folder")". In this case, the "yaml" folder was at path location "/user/home/<user>/Admin/PythonModules", so this was used as the path.